requirements.txt
torch>=2.0.1
torchvision>=0.15.2
timm>=1.0.23
einops>=0.8.1
numpy>=1.24.4
Pillow>=10.4.0
tqdm>=4.67.1

